+++ src/cli.ts
@@ -66,8 +66,8 @@ program
 	.option("--version <version>", "Plugin version; value will be written to the manifest")
 	.option("--force-update-check", "Forces an update check", false)
 	.option("--no-update-check", "Disables updating schemas", true)
+	.option("--no-file-list", "Disables showing the list of files after packaging", false)
 	.option("--ignore-validation", "Bypass validation errors (not recommended)", false)
-	.option("--skip-file-list", "Omits listing the packaged files after packaging", false)
 	.action((path, opts) => pack({ ...opts, path }));
 
 const configCommand = program.command("config").description("Manage the local configuration.");